+++ src/Components/Kite/index.js
@@ -2,13 +2,13 @@ import React from 'react'
 
 import './index.css'
 
-const KiteFlying = () => {
+const Kite = () => {
   return (
     <div className="kite">
-      <div className="tail"></div>
-      <div className="string"></div>
+      <div className="tail" />
+      <div className="string" />
     </div>
   )
 }
 
-export default KiteFlying
+export default Kite
